+++ engine/src/main/java/com/sandboni/core/engine/render/file/json/JSONFileRenderer.java
@@ -3,7 +3,6 @@ package com.sandboni.core.engine.render.file.json;
 import com.sandboni.core.engine.exception.RendererException;
 import com.sandboni.core.engine.render.file.FileOptions;
 import com.sandboni.core.engine.render.file.FileRenderer;
-import com.sandboni.core.engine.render.file.json.util.GsonHelper;
 import com.sandboni.core.engine.render.file.FormatHelper;
 
 import java.util.Collection;
